+++ src/client/components/componentWithSliders/saveButton/saveButton.js
@@ -2,8 +2,8 @@ import React from 'react';
 import './saveButton.css';
 import { saveIcon } from '../../../assets/icons/icons';
 
-const SaveButton = () => (
-  <button type="button" className="SaveButton" id="savePressetes">
+const SaveButton = (props) => (
+  <button type="button" className="SaveButton" onClick={props.onclick}>
     {' '}
     {saveIcon}
     {' '}
