+++ src/client/components/componentWithSliders/savePressetModalBlock/modalSavePressetesWindow/modalSavePressetesWindow.css
