+++ src/client/components/componentWithSliders/savePressetModalBlock/modalSavePressetesWindow/savePressetsToBDAndCancelButton/index.js
