+++ src/client/containers/main/main.js
@@ -5,8 +5,6 @@ import Modes from '../../components/modes/modes';
 import AppHeader from '../../components/appheader/appheader';
 import Equalizer from '../../components/equalizer';
 import ComponentWithSliders from '../../components/componentWithSliders/componentWithSliders';
-import SavePressetModalBlock from '../../components/componentWithSliders/modalSavePressetesWindow';
-
 
 import './main.css';
 
@@ -16,8 +14,7 @@ const Main = () => (
     <AdditionalConfigs />
     <Modes />
     <Equalizer />
-    <ComponentWithSliders />
-    <SavePressetModalBlock />
+    <ComponentWithSliders />    
     <Link to="/login">Login</Link>
     <Link to="/registration">Registration</Link>
   </React.Fragment>
